#24 Move userValidator to middleware folder
